feat: add rect and square drawing
